backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.9  # PostgreSQL sync driver for Alembic
alembic>=1.13.1

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# GitHub Integration
PyGithub>=2.1.1
httpx>=0.26.0  # async HTTP client

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Logging
structlog>=24.1.0

# Utilities
python-frontmatter>=1.1.0  # Markdown frontmatter parsing
python-dotenv>=1.0.0

# Error Tracking (optional, for later)
# sentry-sdk[fastapi]>=1.40.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
